+++ connect/x402/client/src/main/kotlin/org/http4k/connect/x402/X402FacilitatorAction.kt
@@ -1,0 +1,7 @@
+package org.http4k.connect.x402
+
+import dev.forkhandles.result4k.Result
+import org.http4k.connect.Action
+import org.http4k.connect.RemoteFailure
+
+interface X402FacilitatorAction<R> : Action<Result<R, RemoteFailure>>