+++ connect/x402/client/src/main/kotlin/org/http4k/connect/x402/model/FacilitatorRequest.kt
@@ -1,0 +1,9 @@
+package org.http4k.connect.x402.model
+
+import se.ansman.kotshi.JsonSerializable
+
+@JsonSerializable
+data class FacilitatorRequest(
+    val payload: PaymentPayload,
+    val paymentRequirements: PaymentRequirements
+)