+++ connect/x402/fake/src/test/kotlin/org/http4k/connect/x402/FakeX402FacilitatorTest.kt
@@ -1,0 +1,5 @@
+package org.http4k.connect.x402
+
+class FakeX402FacilitatorTest : X402FacilitatorContract {
+    override val http = FakeX402Facilitator()
+}